Address error-prone warnings

+++ src/main/java/org/gaul/s3proxy/crypto/DecryptionInputStream.java
@@ -316,7 +316,7 @@ public class DecryptionInputStream extends FilterInputStream {
         if (n < 0) {
             return 0;
         }
-        ostart += n;
+        ostart += (int) n;
         return n;
     }
